backend/circ-exchange-mvp/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.8.3
pytest==7.4.0
httpx==0.24.0
geopy==2.4.1